// src/interfaces.ts
export interface FacebookMeta {
  appId?: string;
}

export interface LanguageAlternate {
  hrefLang: string;
  href: string;
}

export interface MobileAlternate {
  media: string;
  href: string;
}

export interface OpenGraphArticle {
  authors?: string[];
  expirationTime?: string;
  modifiedTime?: string;
  publishedTime?: string;
  section?: string;
  tags?: string[];
}

export interface OpenGraphBook {
  authors?: string[];
  isbn?: string;
  releaseDate?: string;
  tags?: string[];
}

export interface OpenGraphMedia {
  alt: string;
  height?: number;
  secureUrl?: string;
  type?: string;
  url: string;
  width?: number;
}

export interface OpenGraphMeta {
  article?: OpenGraphArticle;
  book?: OpenGraphBook;
  defaultImageHeight?: number;
  defaultImageWidth?: number;
  description?: string;
  images?: OpenGraphMedia[];
  locale?: string;
  profile?: OpenGraphProfile;
  siteName?: string;
  title?: string;
  type?: string;
  url?: string;
  video?: OpenGraphVideo;
  videos?: OpenGraphMedia[];
}

export interface OpenGraphProfile {
  firstName?: string;
  lastName?: string;
  gender?: string;
  username?: string;
}

export interface OpenGraphVideo {
  actors?: OpenGraphVideoActors[];
  directors?: string[];
  duration?: number;
  releaseDate?: string;
  series?: string;
  tags?: string[];
  writers?: string[];
}
// ...
export interface TwitterMeta {
  /**
   * The card type. Used with all cards.
   */
  card?: TwitterCardType;
  /**
   * The @username of content creator, which may be different than the @username
   * of the site itself. Used with `summary_large_image` cards.
   */
  creator?: string | { id: string };
  /**
   * Description of content (maximum 200 characters). Used with `summary`,
   * `summary_large_image`, and `player` cards.
   */
  description?: string;
  /**
   * The @username of the website. Used with `summary`, `summary_large_image`,
   * `app`, and `player` cards
   */
  site?: string | { id: string };
  /**
   * Title of content (max 70 characters). Used with `summary`, `summary_large_image`, and `player` cards
   */
  title?: string;
  /**
   * The image to use in the card. Images must be less than 5MB in size. JPG,
   * PNG, WEBP and GIF formats are supported. Only the first frame of an
   * animated GIF will be used. SVG is not supported. Used with `summary`,
   * `summary_large_image`, and `player` cards.
   */
  image?: TwitterImageMeta;
  /**
   * The video player to use in the card. Used with the `player` card.
   */
  player?: TwitterPlayerMeta;
  /**
   * Meta used with the `app` card.
   */
  app?: TwitterAppMeta;
}

export type TwitterCardType =
  | "app"
  | "player"
  | "summary"
  | "summary_large_image";

export interface TwitterImageMeta {
  /**
   * The URL of the image to use in the card. This must be an absolute URL,
   * *not* a relative path.
   */
  url: string;
  /**
   * A text description of the image conveying the essential nature of an image
   * to users who are visually impaired. Maximum 420 characters.
   */
  alt: string;
}

export interface TwitterPlayerMeta {
  /**
   * The URL to the player iframe. This must be an absolute URL, *not* a
   * relative path.
   */
  url: string;
  /**
   * The URL to raw video or audio stream. This must be an absolute URL, *not* a
   * relative path.
   */
  stream?: string;
  /**
   * Height of the player iframe in pixels.
   */
  height?: number;
  /**
   * Width of the player iframe in pixels.
   */
  width?: number;
}

export interface TwitterAppMeta {
  name: string | { iPhone?: string; iPad?: string; googlePlay?: string };
  id: { iPhone?: string; iPad?: string; googlePlay?: string };
  url: { iPhone?: string; iPad?: string; googlePlay?: string };
}

export interface SeoConfig {
  bypassTemplate?: boolean;
  canonical?: string;
  defaultTitle?: string;
  description?: string;
  facebook?: FacebookMeta;
  languageAlternates?: LanguageAlternate[];
  mobileAlternate?: MobileAlternate;
  omitGoogleBotMeta?: boolean;
  openGraph?: OpenGraphMeta;
  robots?: RobotsOptions;
  title?: string;
  titleTemplate?: string;
  twitter?: TwitterMeta;
}

// TODO: Use Remix/RR types
export interface RouteArgs {
  data: any;
  parentsData: RouteData;
  params: Params; // Params<string>;
  location: Location;
}
export type RouteData = any;
export type Params<T = any> = any & T;
export type Location = any;

export interface SeoBaseFunction<Return> {
  (config?: SeoConfig): Return;
  (
    config: SeoConfig | ((routeArgs?: RouteArgs) => SeoConfig),
    routeArgs: RouteArgs
  ): Return;
}

export interface SeoFunction
  extends SeoBaseFunction<[V2_HtmlMetaDescriptor[], HtmlLinkDescriptor[]]> {}

export interface SeoMetaFunction
  extends SeoBaseFunction<V2_HtmlMetaDescriptor[]> {}

export interface SeoLinksFunction
  extends SeoBaseFunction<HtmlLinkDescriptor[]> {}